backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.5
bcrypt==4.1.3
passlib==1.7.4
python-jose==3.5.0
PyJWT==2.11.0
python-telegram-bot==22.6
googlemaps==4.10.0
python-multipart==0.0.22
